backend/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cffi==1.17.1
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
fastapi==0.116.0
flatbuffers==25.2.10
fonttools==4.58.5
h11==0.16.0
idna==3.10
jax==0.6.2
jaxlib==0.6.2
kiwisolver==1.4.8
matplotlib==3.10.3
mediapipe==0.10.14
ml_dtypes==0.5.1
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
protobuf==4.25.8
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
scipy==1.16.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
